% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeDEStrand.binMethyl.R
\name{MeDEStrand.binMethyl}
\alias{MeDEStrand.binMethyl}
\title{Function to infer bin-based absolute methylation levels from enrichment signals}
\usage{
MeDEStrand.binMethyl(MSetInput = NULL, CSet = NULL, ccObj = NULL,
  Granges = FALSE)
}
\arguments{
\item{MSetInput}{\linkS4class{MEDIPSset} objects created by function 'MeDEStrand.createSet()', reads mapped to the positive and negative DNA strand are processed separately.}

\item{CSet}{A \linkS4class{COUPLINGset} object.}

\item{ccObj}{Default NULL. Return of internal function call to \code{\link{MeDEStrand.calibrationCurve}}}

\item{Granges}{Default FALSE. Return a vector of inferred absolute methylation levels at user-specified bin size. If TRUE, return a \linkS4class{GRanges} object with each bin's genomic coordinate and its absolute methylation levels.}
}
\value{
Absolute methylation levels in a vector
}
\description{
Function normalizes MeDIP-seq signals by estimating and subsequently correcting CpG bias from the raw signals (i.e. observed bin counts.)
}
\examples{
file.name = "ENCFF002BKV.bam"; BSgenome="BSgenome.Hsapiens.UCSC.hg19"; uniq=1; extend=200; shift=0; ws=100;
chr.select="chr10"
MeDIPSset = MeDEStrand.createSet(file=file.name, BSgenome=BSgenome, extend=extend, shift=shift, uniq=uniq,
window_size=ws, chr.select=chr.select, paired = F)
Bin_methylation = MeDEStrand.binMethyl(MSet = MeDIPSset, CSet = CS )
}
