% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeDEStrand.calibrationCurve.R
\name{MeDEStrand.calibrationCurve}
\alias{MeDEStrand.calibrationCurve}
\title{Function takes a MEDIPSset and finds means of bin counts}
\usage{
MeDEStrand.calibrationCurve(MSet = NULL, CSet = NULL, input = FALSE)
}
\arguments{
\item{MSet}{\linkS4class{MEDIPSset} objects created by function 'MeDEStrand.createSet()'.}

\item{CSet}{A \linkS4class{COUPLINGset} object.}
}
\value{
means of bin counts for each group (categorized by bin CpG counts); estimated upper asymptote for the sigmoidal logistic regression model that estimates CpG bias.
}
\description{
Function takes MEDIPS SET object, re-organize bins into groups of same CpG counts and finds mean bin counts for each group. This is an internal function used by function "MeDEStrand.plotCalibrationCurve()" and function "MeDEStrand.binMethyl()"
}
