% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeDEStrand.cor.R
\name{MeDEStrand.cor}
\alias{MeDEStrand.cor}
\title{Function calculates correlation (Pearson or Spearman) coefficient between inferred bin-based absolute methylation levels with supplied RRBS data}
\usage{
MeDEStrand.cor(MSetInput = NULL, CS = NULL, RRBS = NULL, minRRBS = NULL,
  method = "pearson")
}
\arguments{
\item{MSetInput}{\linkS4class{MEDIPSset} objects.}

\item{RRBS}{A \linkS4class{methylRaw} object ( object returned by function "processBismarkAln()" from package 'MethylKit') or A \linkS4class{GRanges} object with a metadata column 'methylation' to specify the methylation level of each CpG or a '.bed' file with at least these columns: 'chr','start','end','methylation' for CpGs.}

\item{minRRBS}{to filter out bins with RRBS CpGs less than (<) 'minRRBS' number.}

\item{method}{"pearson" or "spearman". To find the Pearson or Spearman correlation coefficient.}

\item{CSet}{A \linkS4class{COUPLINGset} object.}
}
\value{
Pearson/Spearman correlation coefficient value
}
\description{
Function calculates correlation (Pearson or Spearman) coefficient between a vector of inferred bin-based absolute methylation levels with supplied RRBS data. Since RRBS data provides single-resolution CpG cytosine methylation levels, thus RRBS CpGs are binned and the means of CpG methylation in the bins are assigned as the 'true' absolute methylation levels for the bins.
}
\examples{
# Generate MEDIPS SET objects from .bam file
file.name = "ENCFF002BKV.bam"; BSgenome="BSgenome.Hsapiens.UCSC.hg19"; uniq=1; extend=200; shift=0; ws=100;
chr.select="chr10"
MeDIPSset = MeDEStrand.createSet(file=file.name, BSgenome=BSgenome, extend=extend, shift=shift, uniq=uniq,
window_size=ws, chr.select=chr.select, paired = F)
# Load RRBS data (a 'methylRaw' object) saved in the package
fpath <- system.file("data", 'GM12878.RRBS.RData', package="MeDEStrand")
load(fpath)
# Find the Pearson correlation coefficient
correlation = MeDEStrand.cor( MSetInput= MeDIP_single, CS = CS, RRBS = RRBS, minRRBS = 4 , method = "pearson" )
}
