% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeDEStrand.countCG.R
\name{MeDEStrand.countCG}
\alias{MeDEStrand.countCG}
\title{The function counts the frequency of the specified sequence pattern exist in each bin.}
\usage{
MeDEStrand.countCG(pattern = "CG", refObj = NULL)
}
\arguments{
\item{pattern}{defines the sequence pattern, e.g. 'CG' for CpGs.}

\item{refObj}{A \linkS4class{MEDIPSset} object or an object returned by function \code{\link{MeDEStrand.createSet}}}
}
\value{
the counts of specified pattern e.g.'CG' frequencies in bins
}
\description{
The function calculates the local densities(i.e. bin level) of a defined sequence pattern (e.g. 'CG' for CpGs) and returns a \linkS4class{COUPLINGset} object.
}
\examples{
file.name = "ENCFF002BKV.bam"; BSgenome="BSgenome.Hsapiens.UCSC.hg19"; uniq=1; extend=200; shift=0; ws=100;
chr.select="chr10"
MeDIPSset = MeDEStrand.createSet(file=file.name, BSgenome=BSgenome, extend=extend, shift=shift, uniq=uniq,
window_size=ws, chr.select=chr.select,paired = F)
# count CG contents in each bin
CS = MeDEStrand.countCG(pattern="CG", refObj=MeDIPSset)
}
