% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeDEStrand.createSet.R
\name{MeDEStrand.createSet}
\alias{MeDEStrand.createSet}
\title{Function to create a list of two MEDIPS SET objects for reads mapped to the positive and negative DNA strand}
\usage{
MeDEStrand.createSet(file = NULL, extend = 0, shift = 0,
  window_size = 300, BSgenome = NULL, uniq = 0.001, chr.select = NULL,
  paired = F, sample_name = NULL, isSecondaryAlignment = FALSE,
  simpleCigar = TRUE)
}
\arguments{
\item{file}{Path and file name of the input data.}

\item{extend}{defines the number of bases by which the region will be extended before the genome vector is calculated.Regions will be extended along the plus or the minus strand as defined by their provided strand information.}

\item{shift}{As an alternative to the extend parameter, the shift parameter can be specified. Here, the reads are not extended but shifted by the specified number of nucleotides with respect to the given strand infomation.One of the two parameters extend or shift has to be 0.}

\item{window_size}{i.e. bin size. This parameter defines the genomic resolution by which short read coverage is calculated.}

\item{BSgenome}{The reference genome name as defined by BSgenome.}

\item{uniq}{The uniq parameter determines, if all reads mapping to exactly the same genomic position should be kept (uniq = 0), replaced by only one representative (uniq = 1), or if the number of stacked reads should be capped by a maximal number of stacked reads per genomic position determined by a poisson distribution of stacked reads genome wide and by a given p-value (1 > uniq > 0) (deafult: 1e-3). The smaller the p-value, the more reads at the same genomic position are potentially allowed.}

\item{chr.select}{only data at the specified chromosomes will be processed.}

\item{paired}{option for paired end reads.}

\item{sample_name}{name of the sample to be stored with the MEDIPS SET.}

\item{isSecondaryAlignment}{option to import only primary alignments.}

\item{simpleCigar}{option to import only alignments with simple Cigar string.}
}
\value{
\linkS4class{MEDIPSset} objects created for reads mapped to the positive and negative DNA strand respectively
}
\description{
Function creates \linkS4class{MEDIPSset} objects from input data (i.e. aligned short reads) mapped to the positive and negative DNA strand respectively.

Input format:	bam file or tab (|) separated txt file "chr | start | stop  | strand"
}
\examples{
# Set parameters
file.name = "ENCFF002BKV.bam"; BSgenome="BSgenome.Hsapiens.UCSC.hg19"; uniq=1; extend=200; shift=0; ws=100;
chr.select="chr10"
# Generate a list of two MEDIP SET objects from reads mapped to the positive and negative DNA strand
,respectively.
MeDIPSset = MeDEStrand.createSet(file=file.name, BSgenome=BSgenome, extend=extend, shift=shift, uniq=uniq,
window_size=ws, chr.select=chr.select, paired = F)
}
