% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeDEStrand.meth.R
\name{MeDEStrand.meth}
\alias{MeDEStrand.meth}
\title{Function calculates genome wide absolute methylation levels for each provided set and uses t test for differential methylated loci between two groups of MEDIPS SETs (if provided)}
\usage{
MeDEStrand.meth(MSet1 = NULL, MSet2 = NULL, CSet = NULL, chr = NULL,
  p.adj = "bonferroni", minRowSum = 10)
}
\arguments{
\item{MSet1}{first group of MEDIPS SETs. Each group contains several \linkS4class{MEDIPSset} objects.}

\item{MSet2}{second group of MEDIPS SETs. Each group contains several \linkS4class{MEDIPSset} objects. Differential methylated coverage will be calculated by t test and returned, if MSet1 and MSet2 are not empty.}

\item{CSet}{A \linkS4class{COUPLINGset} object. (i.e. corresponding bin CpG density for the provided MSet1 or MSet2). This object is return of function \code{\link{MeDEStrand.countCG}}}

\item{chr}{specifies the chromosome(s) for t test for differential methylated loci between MSet1 and MSet2. Depending on chromosome(s) or all chromosomes are provided, multiple test correction are conducted at chromosome-wide or genome-wide levels. The latter one is more stringent.}

\item{p.adj}{in order to correct p.values for multiple testing, MeDEStrand uses R's 'p.adjust()' function. Therefore, the following methods are available: holm, hochberg, hommel, bonferroni, BH, BY, fdr, none.}

\item{minRowSum}{threshold for a minimum sum of counts across all samples bins (default=10). Bins with lower coverage will not be tested for differential methylation coverage.}
}
\value{
result table of summary statistics and t test results
}
\description{
The function summarizes coverage profiles (bin counts) as well as infers absolute methylation levels for given MEDIPS SETs. In case two groups of MEDIPS SETs are provided at MSet1 and MSet2, the function finds differential methylated loci between MSet1 and MSet2 based on the inferred absolute methylation levels. For each bin/loci, t test is conducted and multiple test correction ( genome-wide/chromosome-wide ) can be specified by user.
}
\examples{
result = MeDEStrand.meth(MSet1=NSCLC_N.MeDIP,  MSet2=NSCLC_T.MeDIP, CSet=CS, chr=c('chr20'),
p.adj="fdr", minRowSum= 12)
}
