% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeDEStrand.plotCalibrationCurve.R
\name{MeDEStrand.plotCalibrationCurve}
\alias{MeDEStrand.plotCalibrationCurve}
\title{Function plots the 'calibration plot', which reveals the CpG density bias}
\usage{
MeDEStrand.plotCalibrationCurve(MSet = NULL, CSet = NULL,
  Forward_Strand = T, plot_chr = "all", main = "Calibration plot",
  xrange = T)
}
\arguments{
\item{MSet}{\linkS4class{MEDIPSset} objects created by function 'MeDEStrand.createSet()' for reads mapped to the positive and negatie DNA strand respectively.}

\item{CSet}{A \linkS4class{COUPLINGset} object. If NULL, it can be calculated and supplied from MSet.}

\item{main}{The title of the calibration plot.}

\item{xrange}{The signal range of the calibration curve typically falls into a low signal range. By setting the xrange parameter to TRUE, the calibration plot will visualize the low signal range only.}

\item{Forward_strand}{if TRUE, plot the 'calibration plot' for the reads mapped to the positive DNA strand(i.e. estimating CpG density bias for reads mapped to the positive DNA strand); if FALSE, plot the 'calibration plot' for the reads mapped to the negative DNA strand (i.e. estimating CpG density bias for reads mapped to the negative DNA strand).}

\item{default}{= "all". It is recommended to call a graphics device (e.g. png("calibrationPlot.png")) before calling the plot command, because R might not be able to plot the full amount of data in reasonable time.}
}
\value{
Calibration plot, i.e. Visualization of the CpG bias curve.
}
\description{
Visualizes the dependency between MeDIP-seq signals and CpG densities by 'calibration curve'. 'calibration curve' estimates CpG bias by fitting a sigmoidal logistic regression curve.
}
\examples{
file.name = "ENCFF002BKV.bam"; BSgenome="BSgenome.Hsapiens.UCSC.hg19"; uniq=1; extend=200; shift=0; ws=100;
chr.select="chr10"
MeDIPSset = MeDEStrand.createSet(file=file.name, BSgenome=BSgenome, extend=extend, shift=shift, uniq=uniq,
window_size=ws, chr.select=chr.select, paired = F)
# plot the 'calibration plot'
MeDEStrand.plotCalibrationCurve( MSet=MeDIPset, CSet=NULL,  Forward_Strand = T, main = NULL, xrange=TRUE)
}
